+++ exp/simd/math_avx512_test.go
@@ -399,7 +399,7 @@ func TestMeanInt16x32(t *testing.T) {
 			got := MeanInt16x32(tc.input)
 			want := int16(0)
 			if len(tc.input) > 0 {
-				want = int16(lo.Sum(tc.input) / int16(len(tc.input)))
+				want = int16(int64(lo.Sum(tc.input)) / int64(len(tc.input)))
 			}
 
 			if got != want {
@@ -434,7 +434,7 @@ func TestMeanInt32x16(t *testing.T) {
 			got := MeanInt32x16(tc.input)
 			want := int32(0)
 			if len(tc.input) > 0 {
-				want = int32(lo.Sum(tc.input) / int32(len(tc.input)))
+				want = int32(int64(lo.Sum(tc.input)) / int64(len(tc.input)))
 			}
 
 			if got != want {
@@ -469,7 +469,7 @@ func TestMeanInt64x8(t *testing.T) {
 			got := MeanInt64x8(tc.input)
 			want := int64(0)
 			if len(tc.input) > 0 {
-				want = int64(lo.Sum(tc.input) / int64(len(tc.input)))
+				want = int64(int64(lo.Sum(tc.input)) / int64(len(tc.input)))
 			}
 
 			if got != want {
@@ -503,7 +503,7 @@ func TestMeanUint8x64(t *testing.T) {
 			got := MeanUint8x64(tc.input)
 			want := uint8(0)
 			if len(tc.input) > 0 {
-				want = uint8(uint8(lo.Sum(tc.input)) / uint8(len(tc.input)))
+				want = uint8(uint64(lo.Sum(tc.input)) / uint64(len(tc.input)))
 			}
 
 			if got != want {
@@ -537,7 +537,7 @@ func TestMeanUint16x32(t *testing.T) {
 			got := MeanUint16x32(tc.input)
 			want := uint16(0)
 			if len(tc.input) > 0 {
-				want = uint16(uint16(lo.Sum(tc.input)) / uint16(len(tc.input)))
+				want = uint16(uint64(lo.Sum(tc.input)) / uint64(len(tc.input)))
 			}
 
 			if got != want {
@@ -571,7 +571,7 @@ func TestMeanUint32x16(t *testing.T) {
 			got := MeanUint32x16(tc.input)
 			want := uint32(0)
 			if len(tc.input) > 0 {
-				want = uint32(uint32(lo.Sum(tc.input)) / uint32(len(tc.input)))
+				want = uint32(uint64(lo.Sum(tc.input)) / uint64(len(tc.input)))
 			}
 
 			if got != want {
@@ -605,7 +605,7 @@ func TestMeanUint64x8(t *testing.T) {
 			got := MeanUint64x8(tc.input)
 			want := uint64(0)
 			if len(tc.input) > 0 {
-				want = uint64(lo.Sum(tc.input) / uint64(len(tc.input)))
+				want = uint64(uint64(lo.Sum(tc.input)) / uint64(len(tc.input)))
 			}
 
 			if got != want {
@@ -640,7 +640,7 @@ func TestMeanFloat32x16(t *testing.T) {
 			got := MeanFloat32x16(tc.input)
 			want := float32(0)
 			if len(tc.input) > 0 {
-				want = float32(lo.Sum(tc.input) / float32(len(tc.input)))
+				want = float32(float64(lo.Sum(tc.input)) / float64(len(tc.input)))
 			}
 
 			const epsilon = 1e-5
@@ -676,7 +676,7 @@ func TestMeanFloat64x8(t *testing.T) {
 			got := MeanFloat64x8(tc.input)
 			want := float64(0)
 			if len(tc.input) > 0 {
-				want = float64(lo.Sum(tc.input) / float64(len(tc.input)))
+				want = lo.Sum(tc.input) / float64(len(tc.input))
 			}
 
 			const epsilon = 1e-10
